Moved optional struct from 'utilities' to 'types'
